+++ src/main/java/codesquad/filter/FilterChain.java
@@ -1,12 +1,22 @@
 package codesquad.filter;
 
+import codesquad.config.FilterConfig;
 import codesquad.http.HttpRequest;
 import codesquad.http.HttpResponse;
 
 /**
  * 필터 체인은 필터의 실행 순서를 제어하는 책임을 가지고 있습니다.
  */
-public interface FilterChain {
+public class FilterChain {
+
+    private final Filter[] filters;
+    private final Router router;
+    private int index = 0;
+
+    public FilterChain(FilterConfig config) {
+        this.filters = config.getFilters();
+        this.router = config.getRouter();
+    }
 
     /**
      * 필터 체인에서 관리되는 필터를 정해진 순서에 따라서 실행하는 메서드입니다.
@@ -14,5 +24,12 @@ public interface FilterChain {
      * @param request  HTTP 요청
      * @param response HTTP 응답
      */
-    void doFilter(HttpRequest request, HttpResponse response);
+    public void doFilter(HttpRequest request, HttpResponse response) {
+        if (index == filters.length) {
+            index++;
+            router.handle(request, response, this);
+        } else if (index < filters.length) {
+            filters[index++].doFilter(request, response, this);
+        }
+    }
 }
